% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RichLab-package.R
\docType{package}
\name{RichLab-package}
\alias{RichLab}
\alias{RichLab-package}
\title{RichLab: Lab Notebooks and Other Templates for the Rich Lab}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package is meant for members of the Rich Lab to have easy access to templates used in our lab notebooks and other tasks. It is also where we store some of our commonly-used helper functions or scripts.
}
\author{
\strong{Maintainer}: Alicia Rich \email{aliciarich@unomaha.edu}

}
\keyword{internal}
